优化一千万的不重复整数的排序

+++ src/main/java/zzz/study/algorithm/sort/BitsMapSort.java
@@ -33,7 +33,7 @@ public class BitsMapSort {
 			}
 			nBitsVector.setBit(arr[i]);		
 		}
-		int bitsLength = nBitsVector.getBitsLength();
+		long bitsLength = nBitsVector.getBitsLength();
 		int count = 0;
 		for (int i=0; i < bitsLength; i++) {
 			if (nBitsVector.testBit(i)) {
